Add logger to gem, improve namespace, add .gitignore file
