+++ .gitignore
@@ -1,0 +1,1 @@
+sidekiq_remove_scheduled-*.gem
