+++ lib/sidekiq_remove_scheduled/binder.rb
@@ -1,17 +1,18 @@
 require 'sidekiq'
 require 'celluloid'
 require 'sidekiq/processor'
-require 'sidekiq_remove_scheduled/job_hook'
-require 'sidekiq_remove_scheduled/process_hook'
+require 'sidekiq_remove_scheduled/logger'
+require 'sidekiq_remove_scheduled/hooks/job_hook'
+require 'sidekiq_remove_scheduled/hooks/process_hook'
 
 module Sidekiq
   module Worker
-    include SidekiqRemoveScheduled::JobHook
+    include SidekiqRemoveScheduled::Hooks::JobHook
   end
 end
 
 module Sidekiq
   class Processor
-    include SidekiqRemoveScheduled::ProcessHook
+    include SidekiqRemoveScheduled::Hooks::ProcessHook
   end
 end
