+++ lib/sidekiq_remove_scheduled/logger.rb
@@ -1,0 +1,15 @@
+module SidekiqRemoveScheduled
+  module Logger
+    class << self
+      attr_accessor :logger
+    end
+    def self.set(object)
+      self.logger = object
+      self
+    end
+    def self.log(msg)
+      return if logger.nil?
+      logger.info(msg)
+    end
+  end
+end
